Make parsing of feedback more robust
Now feedback is assumed to be on a single line that optionally begins with whitespace, then the "-" character,
followed by one or more whitespaces. Previously it was assumed that there always was whitespace before the "-",
i.e. that the feedback was indented. This usually makes things prettier but is no longer enforced by the parsing.